rename exercise history model and schema to category, update queries and mutations accordingly
